% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha_diversity.R
\name{calculate_alpha_diversity}
\alias{calculate_alpha_diversity}
\title{Calculate alpha diversity using multiple rarefaction}
\usage{
calculate_alpha_diversity(
  pobject,
  ntables = 100,
  depth = round(min(sample_sums(pobject)) * 0.9),
  methods = c("Observed", "Chao1", "FaithPD", "Shannon"),
  seedstart = 500,
  verbose = FALSE
)
}
\arguments{
\item{pobject}{A \code{phyloseq} object.}

\item{ntables}{Number or rarefactions to perform. Default 100}

\item{depth}{Rarefaction depth at which the alpha diversity is calculated. Default 90 \% of the lowest sample depth}

\item{methods}{Which alpha diversity metrics to calculate. Default c("Observed","Chao1","FaithPD","Shannon")}

\item{seedstart}{Random seed to use. Default 500}

\item{verbose}{Logical, indicating if the current sample being rarefied should be reported during calculation. Default FALSE}
}
\value{
Data frame with mean and sd alpha diversity for each alpha diveristy metric added to the sample data from phyloseq.
}
\description{
Calculate alpha diversity using multiple rarefaction
}
