% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_taxa.R
\name{collapse_taxa}
\alias{collapse_taxa}
\title{Collapse closely related taxa}
\usage{
collapse_taxa(
  pobject,
  height = 0.05,
  h_method = "average",
  rename = TRUE,
  export_csv = TRUE,
  export_path = "output/clusters.csv",
  verbose = TRUE
)
}
\arguments{
\item{pobject}{Phyloseq object to cluster closely realted species.}

\item{height}{Numeric scalar of the height where the tree should be cut. Default = 0.05}

\item{h_method}{Agglomeration method used, taken from hclust (from stats package). Options "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", or "centroid". Default = "average"}

\item{rename}{Logical, rename the tips the name of the most abundant taxa in the cluster (TRUE), or leave the tip names as the most common taxonomic level (FALSE). Default = TRUE}

\item{export_csv}{Logical, export the csv of species clusters. Default = TRUE}

\item{export_path}{Character, path to export cluster.csv file. Default = "output/cluster.csv"}

\item{verbose}{logical, reports change in the number of taxa as a results of clustering. Default = TRUE}
}
\value{
A phyloseq object with closely related species clustered
}
\description{
A wrapper function that uses tip_glom to group closely related species and/or genus.
}
\examples{
\dontrun{
# Cluster species with a height of 0.2 and don't rename
collapse_taxa(phy, height = 0.2, rename = FALSE)

# Cluster species with a height of 0.01, rename taxa and use the ward.D2 method
collapse_taxa(phy, height = 0.01, method = "ward.D2" )
}

}
