% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_abundance.R
\name{filter_abundance}
\alias{filter_abundance}
\title{Filter by abundance}
\usage{
filter_abundance(pobject, group = NA, min.abundance = 0.01, includes = "any")
}
\arguments{
\item{pobject}{Phyloseq object to filter}

\item{group}{Sample data variable to filter by [DEFAULT = NA]}

\item{min.abundance}{cut-off ratio (taxa with mean abundance below this ratio (0.1 = 10%) will be group in "Others") [DEFAULT = 0.001]}

\item{includes}{If using groups, should taxa be included if they are above the min.abundance in any, all, or a specific number of groups [DEFAULT = "any"]}
}
\value{
A filtered phyloseq object
}
\description{
Filter low abundance taxa in a phyloseq object and group them as "Others". Can be run for all samples together or per group
}
