% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_prevalence.R
\name{filter_prevalence}
\alias{filter_prevalence}
\title{Filter by prevalence}
\usage{
filter_prevalence(
  pobject,
  group = NA,
  min.samples = (length(sample_names(pobject))/10),
  includes = "any"
)
}
\arguments{
\item{pobject}{A \code{phyloseq} object to filter}

\item{group}{Sample data variable to filter by [DEFAILT = NA]}

\item{min.samples}{cut-off value per group (taxa below this value will be group in "Others") [DEFAULT = 10% of samples]}

\item{includes}{If using groups, should taxa be if they are prevalent in more than min.samples in any, all, or a specific number of groups [DEFAULT = "all"]}
}
\value{
A \code{phyloseq} object with the excluded taxa grouped as "Others"
}
\description{
Filter low prevalence taxa in a phyloseq object and group them as "Others". Can be run for all samples or by groups
}
