% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variance.R
\name{filter_variance}
\alias{filter_variance}
\title{Filter by variance}
\usage{
filter_variance(pobject, taxa.lim = 0.9)
}
\arguments{
\item{pobject}{A \code{phyloseq} object to filter}

\item{taxa.lim}{Numeric, if < 1 indicates the ratio of all taxa to keep and if > 1 indicates the number of taxa to keep}
}
\value{
A \code{phyloseq} object with the excluded taxa grouped as "Others"
}
\description{
Filter taxa with low coefficient of variation in a phyloseq object and group them as "Others". Can be run for all samples or by groups
}
