% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sample_alternative.R
\name{merge_samples_alternative}
\alias{merge_samples_alternative}
\title{Merge samples alternative}
\usage{
merge_samples_alternative(x, group, fun = mean)
}
\arguments{
\item{x}{A \code{phyloseq} object to filter}

\item{group}{The variable that should be used to merge the samples}

\item{fun}{Function to be used for merging otu_table. Default mean}
}
\value{
The provided \code{phyloseq} object with an updated otu_table() and sample_data()
}
\description{
Modified version of the phyloseq merge samples function where it is now possible to use the mean instead of sum for otu table.
}
