% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_rarefy.R
\name{multiple_rarefy}
\alias{multiple_rarefy}
\title{Multiple rarefication}
\usage{
multiple_rarefy(
  pobject,
  ntables = 100,
  depth = round(min(sample_sums(pobject)) * 0.9),
  distmethod = "bray",
  summarymeasure = mean,
  seedstart = 500,
  verbose = TRUE
)
}
\arguments{
\item{pobject}{A \code{phyloseq} object.}

\item{ntables}{Number or rarefactions to perform. [DEFAULT = 100]}

\item{depth}{Rarefaction depth at which the alpha diversity is calculated. [DEFAULT = 90 % of the lowest sample depth]}

\item{distmethod}{Distance/dissimilarity metric used to asses iteration similarity. Must be one of: "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao", "mahalanobis", "chisq", "chord", "hellinger", "aitchison", or "robust.aitchison". [DEFAULT = "bray"]}

\item{summarymeasure}{Function to compare iterations. [DEFAULT = mean]}

\item{seedstart}{Random seed to use. [DEFAULT = 500]}

\item{verbose}{Logical, indicating if the current sample being rarefied should be reported during calculation. [DEFAULT = TRUE]}
}
\value{
A rarefied \code{phyloseq} object with rarefied samples
}
\description{
Create multiple rarefications of each sample and select most representative iteration
}
