% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcurve_data.R
\name{Rcurve_data}
\alias{Rcurve_data}
\title{Calculate rarefaction curve data with multiple rarefactions}
\usage{
Rcurve_data(
  pobject,
  ntables = 10,
  step = 250,
  maxdepth = round(unname(quantile(sample_sums(pobject), 0.9))),
  methods = c("Observed", "Chao1", "FaithPD", "Shannon"),
  seedstart = 500,
  verbose = FALSE
)
}
\arguments{
\item{pobject}{A \code{phyloseq} object.}

\item{ntables}{Number or rarefactions to perform. Default 10}

\item{step}{Resolution of sequencing depth (interval between rarefactions). Default 250]}

\item{maxdepth}{Maximum sequencing depth to include Default sample depth of 90th percentile}

\item{methods}{Which alpha diversity metrics to calculate. Default c("Observed","Chao1","FaithPD","Shannon")}

\item{seedstart}{Random seed to use. Default 500}

\item{verbose}{Logical, indicating if information should be reported during calculation. Default FALSE}
}
\value{
Data frame with average alpha diversity per sample at each sequencing depth (from 1 to maxdepth)
}
\description{
Wrapper around calculate_alpha_diversity to generate rarefaction curve data with multiple rarefactions of each sample
}
