% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcurve_data_paral.R
\name{Rcurve_data_paral}
\alias{Rcurve_data_paral}
\title{Parallel Rarefaction Curve}
\usage{
Rcurve_data_paral(
  pobject,
  ntables = 10,
  step = 250,
  maxdepth = 1e+05,
  metrics = c("Observed", "Chao1", "FaithPD", "Shannon"),
  seedstart = 500,
  verbose = FALSE,
  nthreads = 0
)
}
\arguments{
\item{pobject}{A \code{phyloseq} object.}

\item{ntables}{Number or rarefactions to perform. Default 10}

\item{step}{Resolution of sequencing depth (interval between rarefactions). Default 250]}

\item{maxdepth}{Maximum sequencing depth to include Default 100000 - suitable for Nanopore sequencing}

\item{metrics}{Which alpha diversity metrics to calculate. Default c("Observed","Chao1","FaithPD","Shannon")}

\item{seedstart}{Random seed to use. Default 500}

\item{verbose}{Logical, indicating if information should be reported during calculation. Default FALSE}

\item{nthreads}{Integer. Number of parallel threads. 0 = auto detect maxcores available (all cores - 1), 1 = no parallelization, >1 = specific number. Default 0}
}
\value{
dataframe with average alpha diversity per sample at each sequencing depth (from 1 to maxdepth)
}
\description{
Wrapper around calculate_alpha_diversity to generate rarefaction curve data with multiple rarefactions of each sample using parallel processing
}
\examples{
\dontrun{
# Auto detect max cores available (uses all -1)
rdat <- Rcurve_data_paral(phy)

# Use a specific number of cores
rdat <- Rcurve_data_paral(phy, nthreads = 4)

Do not use parallelisation - could also just run Rcurve_data()
rdat <- Rcurve_data_paral(phy, nthreads = 1)
 }
}
