% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_low_abundance.R
\name{remove_low_abundance}
\alias{remove_low_abundance}
\title{Remove low abundance taxa}
\usage{
remove_low_abundance(
  pobject,
  relative = FALSE,
  read_count = 100,
  rel_abun = 1e-04,
  min_prev = 0.01,
  rel_fun = function(x) x/sum(x),
  verbose = TRUE
)
}
\arguments{
\item{pobject}{Phyloseq object to remove low abundance taxa}

\item{relative}{Remove taxa based on relative abundance (TRUE) or raw reads (FALSE). Default = FALSE remove based on raw reads.}

\item{read_count}{Value to filter taxa by before removal when using reads. Default = 100}

\item{rel_abun}{Value to filter taxa by before removal when using relative abundance. Default = 0.0001 (0.01 percent).}

\item{min_prev}{Prevalence (the proportion of samples) that must contain the taxa at the minimum level to be kept. Default = 0.01 (1 percent).}

\item{rel_fun}{Equation used to work out relative abundance, proportion used rather than percent. Default = function(x) x/sum(x).}

\item{verbose}{Logical, prints number of taxa removed. Default = TRUE}
}
\value{
phyloseq object with taxa removed based on rules
}
\description{
Wrapper around phyloseq filter_taxa to allow users to remove (prune) low abundance taxa
}
\examples{
\dontrun{
# Remove taxa where 1\% of the samples does not have more than 100 reads
remove_low_abundance(phy)

# Remove taxa where 10\% of the samples do not contain 0.25 percent relative abundance
remove_low_abundance(phy, relative = TRUE, rel_abun = 0.0025, min_prev = 0.1)

# Remove taxa where 50\% of the samples do not contain 0.001 percent relative abundance
remove_low_abundance(phy, relative = TRUE, rel_abun = 0.0001, min_prev = 0.5)
}

}
