% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rank.R
\name{filter_rank}
\alias{filter_rank}
\title{Filter taxa by rank}
\usage{
filter_rank(pobject, group = NA, min.rank = 10, includes = "any")
}
\arguments{
\item{pobject}{Phyloseq object to filter}

\item{group}{Sample data variable to filter by Default NA}

\item{min.rank}{cut-off ratio (taxa with mean abundance below this ratio (0.1 = 10 \%) will be group in "Others") Default 10}
}
\value{
A filtered phyloseq object
}
\description{
Filter taxa by rank and group low rank taxa as "Others". Can be run for all samples together or per group
}
