% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_GMH_project.R
\name{initiate_GMH_project}
\alias{initiate_GMH_project}
\title{Initiate GMH project}
\usage{
initiate_GMH_project(folders = TRUE, files = TRUE, overwrite = FALSE)
}
\arguments{
\item{folders}{Logical indicator if folder should be created. Default = TRUE}

\item{files}{Logical indicator if Rmarkdown templates should be created. Default = TRUE}

\item{overwrite}{Logical indicator if files should be overwritten if present. Default = FALSE}
}
\value{
Create system files and folders
}
\description{
Creates the folders and/or import the base Rmarkdown templates needed to run basic microbiome analysis using the \code{GMHmicrobiome} pipeline.
}
